Reorganize connection and channel payload modules
