+++ crates/relayer/relayer-components/src/chain/impls/mod.rs
@@ -1,7 +1,6 @@
-pub mod channel_payload;
-pub mod connection_payload;
 pub mod delegate;
 pub mod forward;
+pub mod payload_builders;
 pub mod queries;
 pub mod types;
 pub mod wait_chain_reach_height;
