+++ crates/relayer/relayer-components/src/chain/impls/types/mod.rs
@@ -1,4 +1,3 @@
-pub mod channel_payload;
 pub mod commitment_prefix;
-pub mod connection_payload;
+pub mod payloads;
 pub mod proof;
