+++ crates/relayer/relayer-components/src/chain/types/mod.rs
@@ -1,3 +1,2 @@
 pub mod aliases;
-pub mod channel_payload;
-pub mod connection_payload;
+pub mod payloads;
